requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.27.1
python-dotenv==1.0.1